backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
django-admin==2.0.1
django-cors-headers==3.11.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-rest-framework==0.1.0
django-rest-framework-nested==0.0.1
django-rest-swagger==2.2.0
django-six==1.0.4
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
screen==1.0.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
xlwt==1.3.0
